final modficication for class implementation
